Extract provider usage parsers
